Ensure we pass back helpful return values from calling StatisticsInformation.
The generated class just returns null regardless of what the API
returns, which isn't great. This commit adds an override that returns
the status and body.

While we're adding overrides, might as well add one that strips of the
leading slash from the URI, meaning we don't need to drop that manually
after re-generation. Thus I'm also adding it back to the generated code.